starter_code/backend/requirements.txt
astroid
Click
ecdsa==0.14.1
Flask
Flask-SQLAlchemy
future
isort
itsdangerous
Jinja2
lazy-object-proxy==1.4.3
MarkupSafe
mccabe
pycryptodome
pylint
python-jose
six
SQLAlchemy
typed-ast
Werkzeug
wrapt
Flask-Cors


